+++ src/core/object.h
@@ -11,11 +11,11 @@ class Object {
   public:
     Object() = delete;
 
-    explicit Object(uint32_t id, QPointF position = QPointF(), QPolygonF hitbox = QPolygonF(), float rotationAngle = 0);
+    explicit Object(uint64_t id, QPointF position = QPointF(), QPolygonF hitbox = QPolygonF(), float rotationAngle = 0);
 
     virtual ~Object() = default;
 
-    uint32_t getId() const;
+    uint64_t getId() const;
 
     QPointF getPosition() const;
 
@@ -32,7 +32,7 @@ class Object {
     bool isIntersect(const Object& object) const;
 
   private:
-    uint32_t id;
+    uint64_t id;
     QPointF position;
     QPolygonF hitbox;
     float rotationAngle;
