+++ src/core/resourcebundle.cpp
@@ -3,7 +3,7 @@
 #include "resource.h"
 #include "resourcebundle.h"
 
-core::ResourceBundle::ResourceBundle(uint32_t id, core::ResourceType resourceType, int maxHealth, QPointF position,
+core::ResourceBundle::ResourceBundle(uint64_t id, core::ResourceType resourceType, int maxHealth, QPointF position,
                                      QPolygonF hitbox, float rotationAngle) :
         Object(id, position, std::move(hitbox), rotationAngle), Damageable(maxHealth, maxHealth),
         resourceType(resourceType) {}
