+++ src/core/unit.cpp
@@ -2,7 +2,7 @@
 
 #include "unit.h"
 
-core::Unit::Unit(uint32_t id, QPointF position, QPolygonF hitbox, float rotationAngle,
+core::Unit::Unit(uint64_t id, QPointF position, QPolygonF hitbox, float rotationAngle,
                  const core::Moving& movingProperties, const core::Damaging& damagingProperties,
                  const core::Damageable& damageableProperties)
         : Object(id, position, std::move(hitbox), rotationAngle),
